% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_specificities_discourses.R
\name{calculate_context_specificities}
\alias{calculate_context_specificities}
\title{Calculer les spécificités globales des mots de contexte}
\usage{
calculate_context_specificities(context_tokens_clean, lexique)
}
\arguments{
\item{context_tokens_clean}{Tokens de contexte nettoyés.}

\item{lexique}{Lexique préparé.}
}
\value{
Table des spécificités globales.
}
\description{
Le corpus de référence correspond à l'ensemble des mots
situés autour de tous les mots du lexique, qu'ils soient
simples ou composés de plusieurs mots.
}
\details{
lexicon_word conserve le libellé original en_word.
}
