% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_context_discourses.R
\name{create_context_table}
\alias{create_context_table}
\title{Créer la table détaillée de contexte}
\usage{
create_context_table(context_tokens_clean, pages)
}
\arguments{
\item{context_tokens_clean}{Tokens de contexte nettoyés.}

\item{pages}{Pages préparées.}
}
\value{
Table context.
}
\description{
Créer la table détaillée de contexte
}
