% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_discourses.R
\name{create_info_city}
\alias{create_info_city}
\title{Créer les informations descriptives par ville}
\usage{
create_info_city(pages)
}
\arguments{
\item{pages}{Pages préparées.}
}
\value{
Une ligne par citycode.
}
\description{
Créer les informations descriptives par ville
}
