% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_discourses.R
\name{create_info_city_river}
\alias{create_info_city_river}
\title{Créer les informations descriptives ville-rivière}
\usage{
create_info_city_river(pages)
}
\arguments{
\item{pages}{Pages préparées avec prepare_pages().}
}
\value{
Une ligne par paire citycode-riviere.
}
\description{
Créer les informations descriptives ville-rivière
}
