% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lexicon_tables_discourses.R
\name{create_lexicon_city}
\alias{create_lexicon_city}
\title{Créer la table lexicale par ville}
\usage{
create_lexicon_city(tokens_lexique, pages_city, info_city)
}
\arguments{
\item{tokens_lexique}{Occurrences du lexique.}

\item{pages_city}{Totaux par ville.}

\item{info_city}{Informations descriptives des villes.}
}
\value{
Table lexicon_discourses_city.
}
\description{
tokens_lexique contient une ligne par occurrence de terme
du lexique, y compris pour les expressions multi-mots.
}
