% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lexicon_tables_discourses.R
\name{create_lexicon_city_river}
\alias{create_lexicon_city_river}
\title{Créer la table lexicale par paire ville-rivière}
\usage{
create_lexicon_city_river(tokens_lexique, pages_city_river, info_city_river)
}
\arguments{
\item{tokens_lexique}{Occurrences des mots du lexique.}

\item{pages_city_river}{Totaux de pages analysables.}

\item{info_city_river}{Informations descriptives.}
}
\value{
Table lexicon_discourses_city_river.
}
\description{
tokens_lexique contient maintenant une ligne par occurrence
de terme du lexique, qu'il soit composé d'un ou plusieurs
mots. Une expression comme "climate change" compte donc
comme UNE occurrence.
}
