% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_discourses.R
\name{create_page_totals}
\alias{create_page_totals}
\title{Calculer les volumes de pages analysables}
\usage{
create_page_totals(pages)
}
\arguments{
\item{pages}{Pages préparées.}
}
\value{
Une liste contenant pages_city_river et pages_city.
}
\description{
Calculer les volumes de pages analysables
}
