% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_discourses.R
\name{prepare_lexicon}
\alias{prepare_lexicon}
\title{Préparer le lexique de discours}
\usage{
prepare_lexicon(lexique_brut)
}
\arguments{
\item{lexique_brut}{Table contenant au minimum en_word
et un identifiant fid_word ou fid.}
}
\value{
Un data.frame avec fid_word, en_word et match_word.
}
\description{
Le lexique conserve :
\itemize{
\item en_word : libellé original du terme, utilisé pour l'affichage ;
\item match_word : forme à rechercher dans lemmatext.
}
}
\details{
Si match_word est absent du fichier source, en_word est utilisé
par défaut.
}
