% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_discourses.R
\name{prepare_pages}
\alias{prepare_pages}
\title{Préparer les pages web analysables}
\usage{
prepare_pages(txt_page_work)
}
\arguments{
\item{txt_page_work}{Table source des pages web.}
}
\value{
Table des pages analysables avec page_id.
}
\description{
Seules les pages possédant un lemmatext non vide sont
conservées.
}
