% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_specificities_discourses.R
\name{select_context_specificities}
\alias{select_context_specificities}
\title{Sélectionner les spécificités utilisées par l'application}
\usage{
select_context_specificities(specificities, spec_min = 2, n_min = 5)
}
\arguments{
\item{specificities}{Table issue de
calculate_context_specificities().}

\item{spec_min}{Seuil minimum de spécificité.}

\item{n_min}{Nombre minimum d'occurrences.}
}
\value{
Table filtrée.
}
\description{
Sélectionner les spécificités utilisées par l'application
}
