% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_discourses.R
\name{tokenize_discourses}
\alias{tokenize_discourses}
\title{Tokeniser les textes lemmatisés}
\usage{
tokenize_discourses(pages)
}
\arguments{
\item{pages}{Pages préparées.}
}
\value{
Une ligne par token avec sa position dans la page.
}
\description{
Tokeniser les textes lemmatisés
}
